requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
torch>=2.1
transformers>=4.40
datasets>=2.18
peft>=0.11
evaluate>=0.4
rouge_score>=0.1.2
huggingface_hub>=0.23
nltk
matplotlib
streamlit
plotly
sentencepiece
python-dotenv>=1.0
google-genai>=1.0
